Rudimentary Data Handling
Added the ability to save, load and start game files. Made it possible to fail at loading a file but not crash the game. You can also request the file name and check if the file is in use or not to prevent overwriting existing files. 

Deleted "file_manager.gd" because it was not needed since Godot has its own built-in file handling functions.